Changed Platypus container to host new VMware API Explorer as the UI.
This revision changes the Docker build of the Platypus container so that it
uses the first release of the newly open sourced VMware API Explorer web
component.  The API Explorer provides the ability to browse all Platypus
provided Swagger/Open API docs as well as API docs coming from VMware's
new API services that are in beta at https://code.vmware.com/apis .  The
API explorer provides access to related docs, SDKs, sample code and more as
well.

Note that the features to couple API definitions to related resources does
require some metadata/tags that are missing in the current implementation
in this revision, but these features can be added.

ARCHITECTURE:  What the Dockerfile build now does is a couple of steps:
1) download the current release from https://github.com/vmware/api-explorer
2) extract the built web image
3) use a python based tool to iterate the Platypus included swagger.json
   files and generate a "local.json" metadata file that the API Explorer
   component uses to know what APIs to display.
4) create a minimal size and footprint web container exactly as before.

This revision no longer depends directly on the upstream Swagger UI, but
rather uses it via the API Explorer.

Signed-off-by: Aaron Spear <aspear@vmware.com>